Update PK from int32 to int64 to comply with GCP

+++ frankfund/ServiceLayer/SavingsGoal.cs
@@ -10,7 +10,7 @@ namespace ServiceLayer
 
     public class SavingsGoal
     {
-        public readonly int SGID;
+        public readonly long SGID;
         public string name; 
         public decimal goalAmt;
         public decimal contrAmt;
